+++ packages/bluesky/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.12.2
 
 - Apply automatic generation of moderation sources. ([#1040](https://github.com/myConsciousness/atproto.dart/issues/1040))
+- Add `seenAt` property on `Notifications` object. ([#1046](https://github.com/myConsciousness/atproto.dart/issues/1046))
 
 ## v0.12.1